fixes for task 1C and readme 1

+++ 1_bash/task_C/linux_task_C.sh
@@ -2,4 +2,8 @@
 
 LOGFILE='/var/log/sync.log'
 
-rsync -ah --out-format='%t %o %n' --delete $1 $2  >> $LOGFILE
+if [ $# -eq 0 ]; then
+  echo "Script reques sudo and source and dest as arguments"
+else 
+  rsync -ah --out-format='%t %o %n' --delete $1 $2  >> $LOGFILE
+fi
